+++ packages/owlio_shared/lib/src/constants/rpc_functions.dart
@@ -24,8 +24,6 @@ abstract class RpcFunctions {
   // Cards
   static const buyCardPack = 'buy_card_pack';
   static const openCardPack = 'open_card_pack';
-  static const claimDailyQuestPack = 'claim_daily_quest_pack';
-  static const hasDailyQuestPackClaimed = 'has_daily_quest_pack_claimed';
   static const getDailyQuestProgress = 'get_daily_quest_progress';
   static const claimDailyBonus = 'claim_daily_bonus';
   static const getQuestCompletionStats = 'get_quest_completion_stats';